pyproject.toml
[project]
name = "myprojectllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.47.0",
    "google-generativeai>=0.8.6",
    "gradio>=3.34.0",
]
